refactor(shellhandler): consolidate compositor protocol startup into single-shot init
Replace fragmented shell setup entry points with a single ShellHandler::init(WServer*, WSeat*) that performs one-time initialization for prelaunch splash, app-id resolver, wine window state manager, xdg shell, layer shell, wallpaper shell, and input method helper in a strict order.

Remove redundant nullable/partial initialization branches and enforce single invocation semantics via explicit precondition checks and one-call wiring from Helper after seat readiness, so startup ownership and lifecycle are centralized in ShellHandler.

Also clean related API/header surface by deleting obsolete split init declarations and removing stale includes/forward declarations that were no longer directly used by Helper.
